+++ WebStore/Services/Implementations/CustomerService.cs
@@ -1,3 +1,4 @@
+using WebStore.Exceptions;
 using WebStore.Repositories.Interfaces;
 using WebStore.Services.Interfaces;
 
@@ -11,14 +12,11 @@ namespace WebStore.Services.Implementations
         {
             _customerRepository = customerRepository;
         }
-
-        public bool IsCustomerExists(int customerId)
-        {
-            return _customerRepository.IsCustomerExists(customerId);
-        }
-
         public void DeleteCustomer(int customerId)
         {
+            if (!_customerRepository.IsCustomerExists(customerId))
+                throw new NotFoundException("The customer with such ID is not found.");
+            
             _customerRepository.DeleteCustomer(customerId);
         }
     }
