+++ WebStore/Services/Implementations/ProductService.cs
@@ -2,6 +2,7 @@ using AutoMapper;
 using WebStore.Dto;
 using WebStore.Dto.RequestDtos;
 using WebStore.Dto.UpdateDtos;
+using WebStore.Exceptions;
 using WebStore.Models;
 using WebStore.Repositories.Interfaces;
 using WebStore.Services.Interfaces;
@@ -11,12 +12,15 @@ namespace WebStore.Services.Implementations
     public class ProductService : IProductService
     {
         private readonly IProductRepository _productRepository;
+        private readonly ICategoryRepository _categoryRepository;
         private readonly IMapper _mapper;
 
-        public ProductService(IProductRepository productRepository, IMapper mapper)
+        public ProductService(IProductRepository productRepository, IMapper mapper,
+            ICategoryRepository categoryRepository)
         {
             _productRepository = productRepository;
             _mapper = mapper;
+            _categoryRepository = categoryRepository;
         }
 
         public List<ProductDto> GetProducts()
@@ -26,32 +30,40 @@ namespace WebStore.Services.Implementations
 
         public ProductByIdDto GetProductById(int productId)
         {
+            if (!_productRepository.IsProductExists(productId))
+                throw new NotFoundException("The product with such ID is not found.");
+
             return _mapper.Map<ProductByIdDto>(_productRepository.GetProductById(productId));
         }
 
-        public Product CreateProduct(int categoryId, ProductRequestDto product)
+        public Product CreateProduct(ProductRequestDto product)
         {
-            var newproduct = _mapper.Map<Product>(product);
-            _productRepository.CreateProduct(categoryId, newproduct);
-            return newproduct;
+            if (!_categoryRepository.IsCategoryExists(product.CategoryId))
+                throw new NotFoundException("The category with such ID is not found.");
+
+            var newProduct = _mapper.Map<Product>(product);
+            _productRepository.CreateProduct(newProduct);
+            return newProduct;
         }
 
         public Product UpdateProduct(int productId, ProductUpdateDto product)
         {
-            var updatedproduct = _mapper.Map<Product>(product);
-            _productRepository.UpdateProduct(productId, updatedproduct);
-            updatedproduct.ProductId = productId;
-            return updatedproduct;
+            if (!_productRepository.IsProductExists(productId))
+                throw new NotFoundException("The product with such ID is not found.");
+            if (!_categoryRepository.IsCategoryExists(product.CategoryId))
+                throw new NotFoundException("The category with such ID is not found.");
+            var updatedProduct = _mapper.Map<Product>(product);
+            _productRepository.UpdateProduct(productId, updatedProduct);
+
+            updatedProduct.ProductId = productId;
+            return updatedProduct;
         }
 
         public void DeleteProduct(int productId)
         {
+            if (!_productRepository.IsProductExists(productId))
+                throw new NotFoundException("The product with such ID is not found.");
             _productRepository.DeleteProduct(productId);
-        }
-
-        public bool IsProductExists(int productId)
-        {
-            return _productRepository.IsProductExists(productId);
         }
     }
 }
